requirements.txt
python-dotenv
opencv-python-headless
natsort